[internals] Move `CSPContext` to internals
